DominoEffect

Description

Functions to solve the Domino Effect challenge

Documentation

type Position = (Int, Int, Int)

A (row, column, index) triplet, all starting from 0. Top left is (0, 0, 0), bottom right for maxPips == 6 is (6, 7, 55)

type BonePips = (Pips, Pips)

A Bone has two parts with Pips

type BoneNumber = Int

Every Bone is uniquely numbered

type Bone = (BonePips, BoneNumber)

A Bone is another word for domino piece stone unit / thingy

type Solution = [BoneNumber]

A Solution is a list of BoneNumber corresponding to the Bone of which the Pips match that particular Position in the Puzzle

type Move = (Position, Position, BoneNumber)

A Move as determined in the algorithm, consisting of the BoneNumber to be put on two Positions

dominoEffect :: IO ()

In-module script, for which maxPips should be set

funDominoEffect :: Int -> IO ()

Function to solve the Domino Effect challenge

maxPips :: Int

The maximum number of Pips on a Bone

initialBones :: [Bone]

The initial set of Bones to solve the puzzle with

funInitialBones :: Int -> [Bone]

Function to generate the initial set of Bones, depending on the maximum number of Pips on a Bone

initialPositions :: [Position]

The initial set of Positions in the Puzzle, on which Bones have to be placed

funInitialPositions :: Int -> [Position]

Function to generate the initial set of Positions, depending on the maximum number of Pips on a Bone

initialSolution :: Solution

The initial Solution, consisting entirely of non-existent BoneNumbers

funInitialSolution :: Int -> Solution

Function to generate the initial Solution, depending on the maximum number of Pips on a Bone

defaultBoneNumber :: BoneNumber

The default BoneNumber, not to be used by any Bone in the initial set

maxRow :: Int

The maximum row index

funMaxRow :: Int -> Int

Function to calculate the maximum row index, depending on the maximum number of Pips on a Bone

maxColumn :: Int

The maximum column index

funMaxColumn :: Int -> Int

Function to calculate the maximum column index, depending on the maximum number of Pips on a Bone

numBones :: Int

The number of Bones in the initial set

funNumBones :: Int -> Int

Function to calculate the number of Bones in the initial set, depending on the maximum number of Pips on a Bone

solve :: Puzzle -> [Position] -> [Bone] -> Solution -> [Solution]

The recursive algorithm to solve the Domino Effect challenge

isSolved :: Solution -> Bool

Check if a Solutions solves a Puzzle

okToContinue :: [Position] -> [Bone] -> Bool

Check if continuing is possible with the given Positions and Bones

filterPositions :: [Position] -> Move -> [Position]

Remove Positions on which a Move has been applied

filterBones :: [Bone] -> Move -> [Bone]

Remove Bones that have been used

applyMove :: Solution -> Move -> Solution

Store a Move in the intermediate Solution by updating it

updateList

Arguments

:: [a]
Current list

-> [Position]
Positions to update

-> a
Value to put at the specified positions

-> [a]
Resulting list

Update a list at the specified Positions with the specified value

findMoves

Arguments

:: Puzzle
Puzzle to solve

-> [Position]
Available positions, the head being the one being currently solved

-> [Bone]
Available bones

-> [Move]
Valid positions and bone numbers

Find the Moves that legally place one Bone

bonesWithPips

Arguments

:: Pips
Pips to find on Bones

-> [Bone]
Available Bones

-> [Bone]
Bone matches, with the matched Pips first in the tuple

Find Bones containing the specified Pips

get :: [a] -> Position -> [a]

Get the element at the specified Position, or empty if out of bounds

neighboursInSet :: Position -> [Position] -> [Position]

Get the east (right) and south (down) neighbours of the specified Position from the specified set of Positions

pipsOnBone :: Pips -> Bone -> Bool

Check if the specified Pips are on a Bone